tests/teams/scenarios/mentions/requirements.txt
botbuilder-integration-aiohttp>=4.10.1